Hack cryptmount-setup to run inside script.
Solution is to hardcode the password, which we will ask user
to change later.  We need to run cryptmount-setup inside script
so that we can automatically create a encrypt fs on first-time
bootup.

Note we can not pre-create the encrypted fs because this way everyone
will have the same encryption key.